# backend/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.BoardSummaryResponse:
    properties:
      color:
        type: string
      created_at:
        type: string
      description:
        type: string
      done_cards:
        type: integer
      icon:
        type: string
      id:
        type: string
      last_accessed_at:
        type: string
      members:
        items:
          $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.MemberSummary'
        type: array
      title:
        type: string
      total_cards:
        type: integer
      updated_at:
        type: string
    type: object
...
  github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.ColumnResponse:
    properties:
      cards:
        items:
          $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.CardResponse'
        type: array
      category:
        type: string
      color:
        type: string
      id:
        type: string
      position:
        type: number
      title:
        type: string
    type: object
  github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.CreateBoardRequest:
    properties:
      color:
        type: string
      description:
        description: |-
          Appearance is optional at create time. Omit/null = column default
          (description '', color '#1E40AF', icon 'board'). Same validation as the
          PATCH path so a board can be born with its chosen identity in one request.
        maxLength: 160
        type: string
      icon:
        enum:
        - board
        - rocket
        - target
        - bolt
        - bug
        type: string
      title:
        maxLength: 120
        minLength: 1
        type: string
    required:
    - title
    type: object
...
  github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.MoveCardRequest:
    properties:
      column_id:
        type: string
      position:
        type: number
    required:
    - column_id
    - position
    type: object
...
  github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.ToggleCardDoneRequest:
    properties:
      is_done:
        type: boolean
    required:
    - is_done
    type: object
  github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.UpdateBoardRequest:
    properties:
      budget:
        minimum: 0
        type: number
      color:
        type: string
      description:
        description: |-
          Appearance. Description "" is a valid clear (column is NOT NULL DEFAULT '');
          color must be a hex string; icon is one of the known glyph keys mirrored
          on the client (lib/boardAppearance). omit/null = no change (COALESCE-style
          read-modify-write in the service).
        maxLength: 160
        type: string
      icon:
        enum:
        - board
        - rocket
        - target
        - bolt
        - bug
        type: string
      title:
        maxLength: 120
        minLength: 1
        type: string
    type: object
...
  github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  internal_handler.authUserResponse:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
    type: object
  internal_handler.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  internal_handler.oauthRequest:
    properties:
      email:
        type: string
      full_name:
        maxLength: 120
        minLength: 1
        type: string
      provider:
        enum:
        - google
        - github
        type: string
      provider_id:
        type: string
    required:
    - email
    - full_name
    - provider
    - provider_id
    type: object
  internal_handler.registerRequest:
    properties:
      email:
        type: string
      full_name:
        maxLength: 120
        minLength: 1
        type: string
      password:
        maxLength: 128
        minLength: 8
        type: string
    required:
    - email
    - full_name
    - password
    type: object
  internal_handler.wsTicketResponse:
    properties:
      expires_in:
        type: integer
      ticket:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: Turtask
  description: |-
    Multi-board Kanban + task management with realtime sync.
    Auth uses an HttpOnly `auth_token` cookie issued by /api/auth/login or /api/auth/oauth.
  title: Turtask API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credentials
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_handler.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.authUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
        "401":
          description: invalid credentials
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      summary: Login
      tags:
      - auth
...
  /api/auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.authUserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Current user
      tags:
      - auth
  /api/auth/oauth:
    post:
      consumes:
      - application/json
      parameters:
      - description: Verified provider payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_handler.oauthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.authUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      summary: OAuth callback (programmatic)
      tags:
      - auth
  /api/auth/refresh:
    post:
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      summary: Refresh session
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email, full name, password (min 8)
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_handler.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.authUserResponse'
        "400":
          description: validation error
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
        "409":
          description: email already in use
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      summary: Register
      tags:
      - auth
  /api/boards:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.BoardSummaryResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: List my boards
      tags:
      - boards
    post:
      consumes:
      - application/json
      parameters:
      - description: Board title
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.CreateBoardRequest'
      produces:
      - application/json
      responses:
        "201":
          description: id
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Create board
      tags:
      - boards
  /api/boards/{boardID}:
    delete:
      parameters:
      - description: Board UUID
        in: path
        name: boardID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "403":
          description: not the owner
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Stash board
      tags:
      - boards
    get:
      parameters:
      - description: Board UUID
        in: path
        name: boardID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.ColumnResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
        "404":
          description: not a member or board not found
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Get board contents
      tags:
      - boards
    patch:
      consumes:
      - application/json
      parameters:
      - description: Board UUID
        in: path
        name: boardID
        required: true
        type: string
      - description: Fields to update
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.UpdateBoardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.BoardSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
        "403":
          description: not a manager
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Update board
      tags:
      - boards
...
  /api/cards:
    post:
      consumes:
      - application/json
      parameters:
      - description: Column + title (+ optional fields)
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.CreateCardRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.CardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Create card
      tags:
      - cards
  /api/cards/{cardID}:
    delete:
      parameters:
      - description: Card UUID
        in: path
        name: cardID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Delete card
      tags:
      - cards
    get:
      parameters:
      - description: Card UUID
        in: path
        name: cardID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.CardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Get card detail
      tags:
      - cards
    patch:
      consumes:
      - application/json
      parameters:
      - description: Card UUID
        in: path
        name: cardID
        required: true
        type: string
      - description: Fields to update
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.UpdateCardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.CardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Update card
      tags:
      - cards
  /api/cards/{cardID}/done:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Card UUID
        in: path
        name: cardID
        required: true
        type: string
      - description: Desired done state
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.ToggleCardDoneRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Toggle card done
      tags:
      - cards
  /api/cards/{cardID}/move:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Card UUID
        in: path
        name: cardID
        required: true
        type: string
      - description: Target column + position
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.MoveCardRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Move card
      tags:
      - cards
  /api/columns/{columnID}:
    delete:
      parameters:
      - description: Column UUID
        in: path
        name: columnID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Delete column
      tags:
      - columns
    patch:
      consumes:
      - application/json
      parameters:
      - description: Column UUID
        in: path
        name: columnID
        required: true
        type: string
      - description: Title + category (+ colour)
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.UpdateColumnRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Update column
      tags:
      - columns
  /api/me/settings:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.UserSettingsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: My workspace settings
      tags:
      - me
    patch:
      consumes:
      - application/json
      parameters:
      - description: fields to change
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.UpdateUserSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.UserSettingsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Update my workspace settings
      tags:
      - me
  /api/my-tasks:
    get:
      parameters:
      - description: all|overdue|today|this_week|no_date
        in: query
        name: filter
        type: string
      - description: include unassigned cards on my boards
        in: query
        name: include_unassigned
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.MyWorkResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: My work (cross-board)
      tags:
      - my-tasks
  /api/my-tasks/{cardID}/complete:
    post:
      parameters:
      - description: Card UUID
        in: path
        name: cardID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
        "404":
          description: not the assignee or card missing
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Complete one of my tasks
      tags:
      - my-tasks
  /api/stash:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.StashedBoardDTO'
            type: array
      security:
      - CookieAuth: []
      summary: List stashed boards
      tags:
      - stash
  /api/stash/{boardID}:
    delete:
      parameters:
      - description: Board UUID
        in: path
        name: boardID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Permanently delete stashed board
      tags:
      - stash
...
  /api/ws-ticket:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.wsTicketResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Issue a WebSocket auth ticket
      tags:
      - auth
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: auth_token
    type: apiKey
swagger: "2.0"